Move applicable routes to dwp namespace
