+++ app/controllers/developer_tools/crime_applications_controller.rb
@@ -34,12 +34,12 @@ module DeveloperTools
           edit_steps_client_has_partner_path(crime_application),
           edit_steps_client_details_path(crime_application),
           edit_steps_client_has_nino_path(crime_application),
-          edit_steps_client_benefit_type_path(crime_application),
-          edit_steps_client_benefit_check_result_path(crime_application),
+          edit_steps_dwp_benefit_type_path(crime_application),
+          edit_steps_dwp_benefit_check_result_path(crime_application),
         ]
       )
 
-      redirect_to edit_steps_client_benefit_check_result_path(crime_application)
+      redirect_to edit_steps_dwp_benefit_check_result_path(crime_application)
     end
 
     # rubocop:disable Metrics/AbcSize
