--- app/controllers/steps/client/benefit_exit_controller.rb
@@ -1,7 +1,0 @@
-module Steps
-  module Client
-    class BenefitExitController < Steps::ClientStepController
-      def show; end
-    end
-  end
-end
