--- app/controllers/steps/client/evidence_exit_controller.rb
@@ -1,7 +1,0 @@
-module Steps
-  module Client
-    class EvidenceExitController < Steps::ClientStepController
-      def show; end
-    end
-  end
-end
