+++ app/controllers/steps/dwp/benefit_exit_controller.rb
@@ -1,0 +1,7 @@
+module Steps
+  module DWP
+    class BenefitExitController < Steps::DWPStepController
+      def show; end
+    end
+  end
+end
