+++ app/controllers/steps/dwp/cannot_check_dwp_status_controller.rb
@@ -1,6 +1,6 @@
 module Steps
-  module Client
-    class CannotCheckDWPStatusController < Steps::ClientStepController
+  module DWP
+    class CannotCheckDWPStatusController < Steps::DWPStepController
       include Steps::NoOpAdvanceStep
 
       private
