+++ app/controllers/steps/dwp/evidence_exit_controller.rb
@@ -1,0 +1,7 @@
+module Steps
+  module DWP
+    class EvidenceExitController < Steps::DWPStepController
+      def show; end
+    end
+  end
+end
