+++ app/controllers/steps/dwp/has_benefit_evidence_controller.rb
@@ -1,6 +1,6 @@
 module Steps
-  module Client
-    class HasBenefitEvidenceController < Steps::ClientStepController
+  module DWP
+    class HasBenefitEvidenceController < Steps::DWPStepController
       def edit
         @form_object = HasBenefitEvidenceForm.build(
           current_crime_application
