+++ app/forms/steps/dwp/benefit_type_form.rb
@@ -1,5 +1,5 @@
 module Steps
-  module Client
+  module DWP
     class BenefitTypeForm < Steps::BaseFormObject
       include Steps::HasOneAssociation
       has_one_association :applicant
