+++ app/forms/steps/dwp/cannot_check_benefit_status_form.rb
@@ -1,5 +1,5 @@
 module Steps
-  module Client
+  module DWP
     class CannotCheckBenefitStatusForm < Steps::BaseFormObject
       include Steps::HasOneAssociation
       has_one_association :applicant
