+++ app/forms/steps/submission/cannot_submit_without_nino_form.rb
@@ -1,6 +1,6 @@
 module Steps
   module Submission
-    class CannotSubmitWithoutNinoForm < Steps::Client::CannotCheckBenefitStatusForm
+    class CannotSubmitWithoutNinoForm < Steps::DWP::CannotCheckBenefitStatusForm
       include Steps::HasOneAssociation
       has_one_association :applicant
     end
