+++ app/views/steps/dwp/has_benefit_evidence/edit.html.erb
@@ -11,7 +11,7 @@
                                              legend: { tag: 'h1', size: 'xl', text: t(
                                                '.heading', benefit_type: t(
                                                  current_crime_application.applicant.benefit_type,
-                                                 scope: 'helpers.label.steps_client_benefit_type_form.benefit_type_options'
+                                                 scope: 'helpers.label.steps_dwp_benefit_type_form.benefit_type_options'
                                                )
                                              ) } %>
