+++ config/locales/en/errors.yml
@@ -130,7 +130,7 @@ en:
             nino:
               blank: Enter a National Insurance number
               invalid: Enter a National Insurance number in the correct format
-        steps/client/benefit_type_form:
+        steps/dwp/benefit_type_form:
           attributes:
             benefit_type:
               inclusion: Select a benefit type
@@ -148,11 +148,11 @@ en:
               invalid_year: Enter a valid year
               year_too_early: Date your client last attended an appointment with their JSA work coach is too far in the past
               future_not_allowed: Date your client last attended an appointment with their JSA work coach must be today or in the past
-        steps/client/has_benefit_evidence_form:
+        steps/dwp/has_benefit_evidence_form:
           attributes:
             has_benefit_evidence:
               inclusion: Select yes if you have evidence of your client's benefit
-        steps/client/cannot_check_benefit_status_form: &WILL_ENTER_NINO_VALIDATION
+        steps/dwp/cannot_check_benefit_status_form: &WILL_ENTER_NINO_VALIDATION
           attributes:
             will_enter_nino:
               inclusion: Select whether you will enter the National Insurance number now or later
