+++ config/locales/en/helpers.yml
@@ -61,7 +61,7 @@ en:
         appeal_financial_circumstances_changed: Have your client's financial circumstances changed since the initial application?
       steps_client_appeal_reference_number_form:
         appeal_reference_number: What is the reference number of the original application?
-      steps_client_benefit_type_form:
+      steps_dwp_benefit_type_form:
         benefit_type: Does your client get one of these passporting benefits?
       steps_client_residence_type_form:
         residence_type: Where does your client usually live?
@@ -173,13 +173,13 @@ en:
       steps_client_has_nino_form:
         has_nino: It’s on their National Insurance card, benefit letter, payslip or P60. For example, ‘QQ 12 34 56 C’.
         nino: For example, QQ 12 34 56 C
-      steps_client_benefit_type_form:
+      steps_dwp_benefit_type_form:
         benefit_type: This will help us review your application. It will not affect the DWP check.
         last_jsa_appointment_date: This is sometimes known as ‘signing on’.
       steps_client_cannot_check_benefit_status_form:
         will_enter_nino_options:
           'no': You may not be able to submit your application without the National Insurance number
-      steps_client_has_benefit_evidence_form:
+      steps_dwp_has_benefit_evidence_form:
         has_benefit_evidence: This can be a letter from DWP or JobCentre Plus, or an image from a government app showing your client's name and their benefit details.
         has_benefit_evidence_options:
           'yes': You'll need to upload evidence at the end of this application.
@@ -309,14 +309,6 @@ en:
           appeal_usn: USN
         appeal_maat_id: Enter MAAT ID
         appeal_usn: Enter USN
-      steps_client_benefit_type_form:
-        benefit_type_options:
-          universal_credit: Universal Credit
-          guarantee_pension: Guarantee Credit element of Pension Credit
-          jsa: Income-based Jobseeker's Allowance (JSA)
-          esa: Income-related Employment and Support Allowance (ESA)
-          income_support: Income Support
-          none: My client does not get any of these passporting benefits
       steps_client_residence_type_form:
         residence_type_options:
           rented: In rented accommodation
@@ -328,18 +320,26 @@ en:
           someone_else: In someone else's home
           none: They do not have a fixed home address
         relationship_to_owner_of_usual_home_address: What is your client's relationship to that person?
-      steps_client_has_benefit_evidence_form:
-        has_benefit_evidence_options: *YESNO
-      steps_client_cannot_check_benefit_status_form:
-        will_enter_nino_options:
-          'yes': I will enter the National Insurance number now
-          'no': I will enter the National Insurance number later
       steps_client_contact_details_form:
         telephone_number: What is your client's telephone number? (optional)
         correspondence_address_type_options:
           home_address: Same as usual home address
           providers_office_address: Provider’s office
           other_address: Somewhere else
+      steps_dwp_benefit_type_form:
+        benefit_type_options:
+          universal_credit: Universal Credit
+          guarantee_pension: Guarantee Credit element of Pension Credit
+          jsa: Income-based Jobseeker's Allowance (JSA)
+          esa: Income-related Employment and Support Allowance (ESA)
+          income_support: Income Support
+          none: My client does not get any of these passporting benefits
+      steps_dwp_has_benefit_evidence_form:
+        has_benefit_evidence_options: *YESNO
+      steps_dwp_cannot_check_benefit_status_form:
+        will_enter_nino_options:
+          'yes': I will enter the National Insurance number now
+          'no': I will enter the National Insurance number later
       steps_dwp_confirm_result_form:
         confirm_result_options:
           'yes': 'Yes'
