+++ config/routes.rb
@@ -113,18 +113,18 @@ Rails.application.routes.draw do
         edit_step :date_stamp
         edit_step :has_nino
         show_step :nino_exit
+        edit_step :contact_details
+        edit_step :residence_type
+      end
+
+      namespace :dwp do
         edit_step :benefit_type
         show_step :benefit_exit
         edit_step :benefit_check_result
         edit_step :cannot_check_benefit_status
         edit_step :has_benefit_evidence
         show_step :evidence_exit
-        edit_step :contact_details
         edit_step :cannot_check_dwp_status
-        edit_step :residence_type
-      end
-
-      namespace :dwp do
         edit_step :confirm_result
         edit_step :confirm_details
         show_step :benefit_check_result_exit
