--- spec/controllers/steps/client/evidence_exit_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe Steps::Client::EvidenceExitController, type: :controller do
-  it_behaves_like 'a show step controller'
-end
