+++ spec/controllers/steps/dwp/benefit_check_result_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Steps::DWP::BenefitCheckResultController, type: :controller do
+  it_behaves_like 'a no-op advance step controller', :benefit_check_result, Decisions::DWPDecisionTree
+end
