+++ spec/controllers/steps/dwp/benefit_type_controller_spec.rb
@@ -1,0 +1,6 @@
+require 'rails_helper'
+
+RSpec.describe Steps::DWP::BenefitTypeController, type: :controller do
+  it_behaves_like 'a generic step controller', Steps::DWP::BenefitTypeForm, Decisions::DWPDecisionTree
+  it_behaves_like 'a step that can be drafted', Steps::DWP::BenefitTypeForm
+end
