+++ spec/controllers/steps/dwp/evidence_exit_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Steps::DWP::EvidenceExitController, type: :controller do
+  it_behaves_like 'a show step controller'
+end
