+++ spec/forms/steps/dwp/has_benefit_evidence_form_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Steps::Client::CannotCheckBenefitStatusForm do
+RSpec.describe Steps::DWP::HasBenefitEvidenceForm do
   subject { described_class.new(arguments) }
 
   let(:arguments) do
@@ -13,12 +13,12 @@ RSpec.describe Steps::Client::CannotCheckBenefitStatusForm do
   end
 
   let(:form_attributes) do
-    { will_enter_nino: }
+    { has_benefit_evidence: }
   end
 
   let(:crime_application) { instance_double(CrimeApplication, applicant: applicant_record) }
   let(:applicant_record) { Applicant.new }
-  let(:will_enter_nino) { 'yes' }
+  let(:has_benefit_evidence) { 'yes' }
 
   describe '#choices' do
     it 'returns the possible choices' do
@@ -29,8 +29,8 @@ RSpec.describe Steps::Client::CannotCheckBenefitStatusForm do
   end
 
   describe '#save' do
-    context 'when `will_enter_nino` is not provided' do
-      let(:will_enter_nino) { '' }
+    context 'when `has_benefit_evidence` is not provided' do
+      let(:has_benefit_evidence) { '' }
 
       it 'returns false' do
         expect(subject.save).to be(false)
@@ -38,12 +38,12 @@ RSpec.describe Steps::Client::CannotCheckBenefitStatusForm do
 
       it 'has a validation error on the field' do
         expect(subject).not_to be_valid
-        expect(subject.errors.of_kind?(:will_enter_nino, :inclusion)).to be(true)
+        expect(subject.errors.of_kind?(:has_benefit_evidence, :inclusion)).to be(true)
       end
     end
 
-    context 'when `will_enter_nino` is not valid' do
-      let(:will_enter_nino) { 'maybe' }
+    context 'when `has_benefit_evidence` is not valid' do
+      let(:has_benefit_evidence) { 'maybe' }
 
       it 'returns false' do
         expect(subject.save).to be(false)
@@ -51,14 +51,14 @@ RSpec.describe Steps::Client::CannotCheckBenefitStatusForm do
 
       it 'has a validation error on the field' do
         expect(subject).not_to be_valid
-        expect(subject.errors.of_kind?(:will_enter_nino, :inclusion)).to be(true)
+        expect(subject.errors.of_kind?(:has_benefit_evidence, :inclusion)).to be(true)
       end
     end
 
     context 'when form is valid' do
       it 'saves the record' do
         expect(applicant_record).to receive(:update).with(
-          { 'will_enter_nino' => YesNoAnswer::YES }
+          { 'has_benefit_evidence' => YesNoAnswer::YES }
         ).and_return(true)
 
         expect(subject.save).to be(true)
