+++ spec/services/decisions/client_decision_tree_spec.rb
@@ -300,7 +300,7 @@ RSpec.describe Decisions::ClientDecisionTree do
     let(:step_name) { :has_nino }
 
     context 'when the application is means tested' do
-      it { is_expected.to have_destination(:benefit_type, :edit, id: crime_application) }
+      it { is_expected.to have_destination('/steps/dwp/benefit_type', :edit, id: crime_application) }
     end
 
     context 'when the application is not means tested' do
@@ -308,170 +308,6 @@ RSpec.describe Decisions::ClientDecisionTree do
 
       it { is_expected.to have_destination('/steps/case/urn', :edit, id: crime_application) }
     end
-  end
-
-  # rubocop:disable RSpec/MultipleMemoizedHelpers
-  context 'when the step is `benefit_type`' do
-    let(:form_object) { double('FormObject', benefit_type:) }
-    let(:applicant_double) { double(Applicant, has_nino:) }
-    let(:step_name) { :benefit_type }
-    let(:benefit_type) { BenefitType::UNIVERSAL_CREDIT }
-    let(:has_nino) { YesNoAnswer::YES }
-    let(:feature_flag_means_journey_enabled) { false }
-
-    before do
-      allow(crime_application).to receive_messages(applicant: applicant_double,
-                                                   benefit_check_passported?: benefit_check_passported)
-
-      allow(applicant_double).to receive_messages(passporting_benefit:)
-
-      allow(DWP::UpdateBenefitCheckResultService).to receive(:call).with(applicant_double).and_return(true)
-
-      allow(FeatureFlags).to receive(:means_journey) {
-        instance_double(FeatureFlags::EnabledFeature, enabled?: feature_flag_means_journey_enabled)
-      }
-    end
-
-    context 'and the benefit type is `none`' do
-      let(:feature_flag_means_journey_enabled) { false }
-      let(:benefit_type) { BenefitType::NONE }
-      let(:benefit_check_passported) { false }
-      let(:passporting_benefit) { nil }
-
-      it { is_expected.to have_destination(:benefit_exit, :show, id: crime_application) }
-
-      context 'and feature flag `means_journey` is enabled' do
-        let(:feature_flag_means_journey_enabled) { true }
-
-        it { is_expected.to have_destination('/steps/case/urn', :edit, id: crime_application) }
-      end
-    end
-
-    context 'when application has been already passported on benefit check' do
-      let(:benefit_check_passported) { true }
-      let(:passporting_benefit) { nil }
-
-      it { is_expected.to have_destination(:benefit_check_result, :edit, id: crime_application) }
-    end
-
-    context 'when application does not have an existing passporting on benefit check' do
-      let(:benefit_check_passported) { false }
-
-      context 'when the applicant has a passporting benefit' do
-        context 'has correct next step' do
-          let(:passporting_benefit) { true }
-
-          it { is_expected.to have_destination(:benefit_check_result, :edit, id: crime_application) }
-        end
-      end
-
-      context 'when the applicant does not have a passporting benefit' do
-        context 'has correct next step' do
-          let(:passporting_benefit) { false }
-
-          it { is_expected.to have_destination('steps/dwp/confirm_result', :edit, id: crime_application) }
-        end
-      end
-
-      context 'when the benefit checker cannot check on the status of the passporting benefit' do
-        context 'has correct next step' do
-          let(:passporting_benefit) { nil }
-
-          it { is_expected.to have_destination(:cannot_check_dwp_status, :edit, id: crime_application) }
-        end
-      end
-    end
-
-    context 'when the applicant does not have a nino' do
-      let(:passporting_benefit) { nil }
-      let(:benefit_check_passported) { false }
-      let(:has_nino) { YesNoAnswer::NO }
-
-      it { is_expected.to have_destination(:cannot_check_dwp_status, :edit, id: crime_application) }
-
-      context 'and feature flag `means_journey` is enabled' do
-        let(:feature_flag_means_journey_enabled) { true }
-
-        it { is_expected.to have_destination(:cannot_check_benefit_status, :edit, id: crime_application) }
-      end
-    end
-  end
-  # rubocop:enable RSpec/MultipleMemoizedHelpers
-
-  context 'when the step is `has_benefit_evidence`' do
-    let(:form_object) { double('FormObject', applicant:, has_benefit_evidence:) }
-    let(:step_name) { :has_benefit_evidence }
-    let(:feature_flag_means_journey_enabled) { true }
-
-    before do
-      allow(FeatureFlags).to receive(:means_journey) {
-        instance_double(FeatureFlags::EnabledFeature, enabled?: feature_flag_means_journey_enabled)
-      }
-    end
-
-    context 'and the answer is `yes`' do
-      let(:has_benefit_evidence) { YesNoAnswer::YES }
-
-      it { is_expected.to have_destination('/steps/case/urn', :edit, id: crime_application) }
-    end
-
-    context 'and the answer is `no`' do
-      let(:has_benefit_evidence) { YesNoAnswer::NO }
-
-      it { is_expected.to have_destination('/steps/case/urn', :edit, id: crime_application) }
-    end
-
-    context 'and the answer is `no` for a non means tested application' do
-      let(:feature_flag_means_journey_enabled) { false }
-      let(:has_benefit_evidence) { YesNoAnswer::NO }
-
-      it { is_expected.to have_destination(:evidence_exit, :show, id: crime_application) }
-    end
-  end
-
-  context 'when the step is `benefit_check_result`' do
-    let(:form_object) { double('FormObject') }
-    let(:step_name) { :benefit_check_result }
-
-    it { is_expected.to have_destination('/steps/case/urn', :edit, id: crime_application) }
-  end
-
-  context 'when the step is `after_cannot_check_benefit_status`' do
-    let(:form_object) { double('FormObject', applicant:, will_enter_nino:) }
-    let(:step_name) { :cannot_check_benefit_status }
-
-    context 'and the answer is `yes`' do
-      let(:will_enter_nino) { YesNoAnswer::YES }
-
-      it { is_expected.to have_destination(:has_nino, :edit, id: crime_application) }
-    end
-
-    context 'and the answer is `no`' do
-      let(:will_enter_nino) { YesNoAnswer::NO }
-
-      it { is_expected.to have_destination('/steps/case/urn', :edit, id: crime_application) }
-    end
-  end
-
-  context 'when the step is `cannot_check_dwp_status`' do
-    # The dwp check decision tree has been tested for the benefit type step so this does not test all routes
-
-    let(:form_object) { double('FormObject') }
-    let(:step_name) { :cannot_check_dwp_status }
-    let(:benefit_check_passported) { false }
-    let(:passporting_benefit) { false }
-    let(:applicant_double) { double(Applicant) }
-
-    before do
-      allow(crime_application).to receive_messages(applicant: applicant_double,
-                                                   benefit_check_passported?: benefit_check_passported)
-
-      allow(applicant_double).to receive_messages(passporting_benefit:)
-
-      allow(DWP::UpdateBenefitCheckResultService).to receive(:call).with(applicant_double).and_return(true)
-    end
-
-    it { is_expected.to have_destination('steps/dwp/confirm_result', :edit, id: crime_application) }
   end
 
   context 'when the step is `residence_type`' do
